Enemy State Change
괴인 상태변경 동작 확인
