--- CRAZYMAN/Assets/Scripts/Multi/NetworkPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d5979841f889cd54c9aed0e21bd78b29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
